Refactor FavouriteList and Feed components to enhance loading state management
- Integrated `Suspense` and `withFallback` into `FavouriteList`, `Content`, and `Item` components for improved loading experiences.
- Removed redundant `Pending` and `Data` components to streamline the structure and enhance maintainability.
- Updated various components to utilize colocated fallback UIs, ensuring consistent loading states across the application.
- Added `data-action` attributes for better user interaction feedback and accessibility.
