+++ apps/app/src/app/@buyer/favourite/FavouriteListPage/FavouriteList/Content.tsx
@@ -1,8 +1,7 @@
 import type { MarkSuspense } from "@use-pico/client/type";
 import { VisibleContainer } from "@use-pico/client/ui/container";
-import { type FC, useCallback } from "react";
+import { Suspense, type FC, useCallback } from "react";
 import { Item } from "./Item";
-import { Pending } from "./Item/Pending";
 
 export namespace Content {
 	export interface Props extends MarkSuspense.Props {
@@ -11,7 +10,7 @@ export namespace Content {
 }
 
 export const Content: FC<Content.Props> = ({ _suspense, listingIds }) => {
-	const placeholder = useCallback(() => <Pending />, []);
+	const placeholder = useCallback(() => <Item.Fallback />, []);
 
 	return (
 		<>
@@ -22,7 +21,12 @@ export const Content: FC<Content.Props> = ({ _suspense, listingIds }) => {
 						id={listingId}
 						placeholder={placeholder}
 					>
-						<Item listingId={listingId} />
+						<Suspense fallback={<Item.Fallback />}>
+							<Item
+								_suspense={"I know"}
+								listingId={listingId}
+							/>
+						</Suspense>
 					</VisibleContainer>
 				);
 			})}
