--- apps/app/src/app/@buyer/favourite/FavouriteListPage/FavouriteList/Data/index.ts
@@ -1,1 +1,0 @@
-export { Data } from "./Data";
