+++ apps/app/src/app/@buyer/favourite/FavouriteListPage/FavouriteListPage.tsx
@@ -1,6 +1,6 @@
 import { translator } from "@use-pico/common/translator";
 import { TitleContainer } from "@zbav-se.me/ui/container";
-import type { FC } from "react";
+import { type FC, Suspense } from "react";
 import { BackHomeButton } from "~/app/@common/nav/BackHomeButton";
 import { HomeMenuButton } from "~/app/@user/home/~public/HomeMenuButton";
 import { FavouriteList } from "./FavouriteList";
@@ -17,7 +17,9 @@ export const FavouriteListPage: FC<FavouriteListPage.Props> = (props) => {
 			right={<HomeMenuButton />}
 			{...props}
 		>
-			<FavouriteList />
+			<Suspense fallback={<FavouriteList.Fallback />}>
+				<FavouriteList _suspense={"I know"} />
+			</Suspense>
 		</TitleContainer>
 	);
 };
