+++ apps/app/src/app/@buyer/feed/FeedEditor/FeedEditorSheet.tsx
@@ -2,11 +2,11 @@ import { BottomSheet } from "@use-pico/client/ui/bottom-sheet";
 import { translator } from "@use-pico/common/translator";
 import type { StateType } from "@use-pico/common/type";
 import { CloseButton } from "@zbav-se.me/ui/button";
-import type { FC } from "react";
+import { type FC, Suspense } from "react";
 import { FeedEditor } from "./FeedEditor";
 
 export namespace FeedEditorSheet {
-	export interface Props extends FeedEditor.Props {
+	export interface Props extends Omit<FeedEditor.Props, "_suspense"> {
 		state: StateType.Simple<boolean>;
 	}
 }
@@ -33,7 +33,12 @@ export const FeedEditorSheet: FC<FeedEditorSheet.Props> = ({ state, ...props }) 
 				),
 			})}
 		>
-			<FeedEditor {...props} />
+			<Suspense fallback={<FeedEditor.Fallback />}>
+				<FeedEditor
+					{...props}
+					_suspense={"I know"}
+				/>
+			</Suspense>
 		</BottomSheet>
 	);
 };
