+++ apps/app/src/app/@buyer/feed/FeedListPage/FeedList/FeedList.tsx
@@ -3,7 +3,7 @@ import { Container, SpinnerContainer } from "@use-pico/client/ui/container";
 import { EmptyState } from "@use-pico/client/ui/empty-state";
 import { withFallback } from "@use-pico/client/utils";
 import { withFeedQuery } from "@zbav-se.me/sdk/query/buyer/feed";
-import { useMemo } from "react";
+import { Suspense, useMemo } from "react";
 import { CreateButton } from "./CreateButton";
 import { Empty } from "./Empty";
 import { Item } from "./Item";
@@ -60,10 +60,15 @@ export const FeedList = withFallback(({ _suspense, ...props }: FeedList.Props) =
 			<EmptyState check={check}>
 				{feedCollection.map((feedId) => {
 					return (
-						<Item
+						<Suspense
 							key={feedId}
-							feedId={feedId}
-						/>
+							fallback={<Item.Fallback />}
+						>
+							<Item
+								_suspense={"I know"}
+								feedId={feedId}
+							/>
+						</Suspense>
 					);
 				})}
