+++ apps/app/src/app/@buyer/feed/FeedListingPage/Content/index.ts
@@ -1,0 +1,1 @@
+export { Content } from "./Content";
