+++ apps/app/src/app/@buyer/feed/FeedListingPage/FavouriteButton/index.ts
@@ -1,0 +1,1 @@
+export { FavouriteButton } from "./FavouriteButton";
