+++ apps/app/src/app/@buyer/feed/FeedListingPage/FlagButton/index.ts
@@ -1,0 +1,1 @@
+export { FlagButton } from "./FlagButton";
