+++ apps/app/src/app/@buyer/feed/FeedListingPage/IgnoreButton/index.ts
@@ -1,0 +1,1 @@
+export { IgnoreButton } from "./IgnoreButton";
