+++ apps/app/src/app/@buyer/feed/FeedListingPage/ListingCard/index.ts
@@ -1,0 +1,1 @@
+export { ListingCard } from "./ListingCard";
