+++ apps/app/src/app/@buyer/feed/FeedListingPage/ListingCard/section/HeroSection.tsx
@@ -1,11 +1,11 @@
 import { Container } from "@use-pico/client/ui/container";
 import type { tListing } from "@zbav-se.me/sdk/api/buyer";
 import { HeroImage } from "@zbav-se.me/ui/img";
-import type { FC } from "react";
+import { type FC, Suspense } from "react";
 import { useUpload } from "~/app/@common/gallery/hook/useUpload";
 import { ListingPrice } from "~/app/@common/listing/ui/ListingPrice";
 import { LocationBadge } from "~/app/@common/location/ui/LocationBadge";
-import { FavouriteButton } from "../../FavouriteButton/FavouriteButton";
+import { FavouriteButton } from "../../FavouriteButton";
 import { TransactionButton } from "../../TransactionButton/TransactionButton";
 
 export namespace HeroSection {
@@ -49,26 +49,52 @@ export const HeroSection: FC<HeroSection.Props> = ({ feedId, listing, onView }) 
 				/>
 
 				{listing.my ? null : (
-					<FavouriteButton
-						feedId={feedId}
-						listingId={listing.id}
-						iconProps={{
-							ui: {
-								text: "xl",
-							},
-						}}
-						ui={{
-							tone: "secondary",
-							theme: "light",
-							round: "full",
-							square: "md",
-							justify: "center",
-							items: "center",
-							size: undefined,
-							inner: undefined,
-							snapTo: "top-right",
-						}}
-					/>
+					<Suspense
+						fallback={
+							<FavouriteButton.Fallback
+								feedId={feedId}
+								listingId={listing.id}
+								iconProps={{
+									ui: {
+										text: "xl",
+									},
+								}}
+								ui={{
+									tone: "secondary",
+									theme: "light",
+									round: "full",
+									square: "md",
+									justify: "center",
+									items: "center",
+									size: undefined,
+									inner: undefined,
+									snapTo: "top-right",
+								}}
+							/>
+						}
+					>
+						<FavouriteButton
+							_suspense={"I know"}
+							feedId={feedId}
+							listingId={listing.id}
+							iconProps={{
+								ui: {
+									text: "xl",
+								},
+							}}
+							ui={{
+								tone: "secondary",
+								theme: "light",
+								round: "full",
+								square: "md",
+								justify: "center",
+								items: "center",
+								size: undefined,
+								inner: undefined,
+								snapTo: "top-right",
+							}}
+						/>
+					</Suspense>
 				)}
 
 				<HeroImage
