+++ apps/app/src/app/@buyer/feed/FeedListingPage/SellerInfo/SellerInfo.tsx
@@ -1,21 +1,46 @@
-import { type FC, Suspense } from "react";
-import { Data } from "./Data";
-import { Pending } from "./Pending";
+import { Icon, ShowIcon } from "@use-pico/client/icon";
+import type { MarkSuspense } from "@use-pico/client/type";
+import { LabelValue } from "@use-pico/client/ui/container";
+import { withFallback } from "@use-pico/client/utils";
+import { translator } from "@use-pico/common/translator";
+import { withListingSellerInfoQuery } from "@zbav-se.me/sdk/query/buyer/listing";
+import { RatingIcon } from "~/app/@common/score/ui/RatingIcon";
 
 export namespace SellerInfo {
-	export interface Props extends Omit<Data.Props, "_suspense"> {
-		//
+	export interface Props extends MarkSuspense.Props {
+		listingId: string;
+		onView(view: "seller-info"): void;
 	}
 }
 
-export const SellerInfo: FC<SellerInfo.Props> = (props) => {
-	return (
-		<Suspense fallback={<Pending />}>
-			<Data
-				_suspense={"I know"}
+export const SellerInfo = withFallback(
+	({ _suspense, listingId, onView }: SellerInfo.Props) => {
+		const { data: sellerInfo } = withListingSellerInfoQuery.useSuspenseQuery({
+			listingId,
+		});
+
+		return (
+			<LabelValue
 				data-ui={"SellerInfo"}
+				data-action={"open seller info"}
+				textLabel={translator.text("Listing seller hint (label)")}
+				textValue={
+					sellerInfo.events ? <RatingIcon rating={sellerInfo.events.score.rank} /> : null
+				}
+				textEmpty={translator.text("Listing seller info not available (empty)")}
+				action={<Icon icon={ShowIcon} />}
+				onClick={() => onView("seller-info")}
+			/>
+		);
+	},
+	(props: Omit<SellerInfo.Props, "_suspense">) => {
+		return (
+			<LabelValue
+				textLabel={translator.text("Listing seller hint (label)")}
+				textValue={null}
+				action={<Icon icon={ShowIcon} />}
 				{...props}
 			/>
-		</Suspense>
-	);
-};
+		);
+	},
+);
