+++ apps/app/src/app/@buyer/feed/FeedListingPage/SellerInfo/index.ts
@@ -1,0 +1,1 @@
+export { SellerInfo } from "./SellerInfo";
