+++ apps/app/src/app/@buyer/feed/FeedListingPage/ThumbDislikeButton/ThumbDislikeButton.tsx
@@ -1,21 +1,79 @@
-import { type FC, Suspense } from "react";
-import { Data } from "./Data";
-import { Pending } from "./Pending";
+import { DislikeIcon } from "@use-pico/client/icon";
+import type { MarkSuspense } from "@use-pico/client/type";
+import { Button } from "@use-pico/client/ui/button";
+import { Tx } from "@use-pico/client/ui/tx";
+import { withFallback } from "@use-pico/client/utils";
+import { withThumbCreateMutation } from "@zbav-se.me/sdk/mutation/buyer/thumb";
+import { withListingQuery } from "@zbav-se.me/sdk/query/buyer/listing";
 
 export namespace ThumbDislikeButton {
-	export interface Props extends Omit<Data.Props, "_suspense"> {
-		//
+	export interface Props extends Button.Props, MarkSuspense.Props {
+		listingId: string;
 	}
 }
 
-export const ThumbDislikeButton: FC<ThumbDislikeButton.Props> = (props) => {
-	return (
-		<Suspense fallback={<Pending />}>
-			<Data
-				_suspense={"I know"}
+export const ThumbDislikeButton = withFallback(
+	({ _suspense, listingId, ui, ...props }: ThumbDislikeButton.Props) => {
+		const { data: listing } = withListingQuery.useFetchQuery(listingId);
+		const update = withListingQuery.useUpdate();
+		const thumbCreateMutation = withThumbCreateMutation.useMutation({
+			onSuccess: update,
+			meta: {
+				mutationId: listing.id,
+			},
+		});
+		const isMutating = withThumbCreateMutation.useIsMutating({
+			mutationId: listing.id,
+		});
+
+		const hasThumb = listing.thumb !== null;
+		const isDisliked = listing.thumb === "dislike";
+
+		return (
+			<Button
 				data-ui={"ThumbDislikeButton"}
+				data-action={"dislike listing"}
+				iconEnabled={DislikeIcon}
+				iconProps={{
+					ui: {
+						text: "xl",
+					},
+				}}
+				disabled={hasThumb || isMutating}
+				loading={thumbCreateMutation.isPending}
+				onClick={() => {
+					thumbCreateMutation.mutate({
+						listingId: listing.id,
+						type: "dislike",
+					});
+				}}
+				ui={{
+					tone: isDisliked ? "secondary" : "neutral",
+					theme: "light",
+					size: "default",
+					justify: "start",
+					...ui,
+				}}
 				{...props}
 			/>
-		</Suspense>
-	);
-};
+		);
+	},
+	({ ui, ...props }: Omit<ThumbDislikeButton.Props, "_suspense">) => {
+		return (
+			<Button
+				disabled
+				loading
+				ui={{
+					tone: "secondary",
+					theme: "light",
+					size: "default",
+					justify: "start",
+					...ui,
+				}}
+				{...props}
+			>
+				<Tx label="Loading... (button)" />
+			</Button>
+		);
+	},
+);
