+++ apps/app/src/app/@buyer/feed/FeedListingPage/ThumbDislikeButton/index.ts
@@ -1,0 +1,1 @@
+export { ThumbDislikeButton } from "./ThumbDislikeButton";
