+++ apps/app/src/app/@buyer/feed/FeedListingPage/ThumbLikeButton/index.ts
@@ -1,0 +1,1 @@
+export { ThumbLikeButton } from "./ThumbLikeButton";
